moved main component pages to container directory
