+++ frontend/src/containers/HomePage.tsx
@@ -1,10 +1,10 @@
 import React from "react";
 
-import About from "../about/About";
 import Footer from "../components/Footer";
 import Hero from "../components/Hero";
-import Contact from "../contact/Contact";
-import Projects from "../projects/Projects";
+import About from "./about/About";
+import Contact from "./contact/Contact";
+import Projects from "./projects/Projects";
 
 const HomePage = () => {
   return (
