+++ frontend/src/containers/about/components/BorderLinearProgress.tsx
